feat(FOSMVVM)!: generated Localizable overload surface replaces hand-written mirrors
Replaces the hand-written Localizable view overloads (Label, LabeledContent,
Tab, TextField, ContentUnavailableView, Text) with the generator's emitted
surface under Generated/, relocates shared resolution machinery into
LocalizableViews.swift, and wires the generated files into Package.swift.

BREAKING CHANGES:
- defaultTitle: parameter renamed to defaultValue:
- existential `any` parameters narrowed to `some`
- ContentUnavailableView defaultValue argument reordered
