--- Sources/FOSMVVM/SwiftUI Support/ContentUnavailableView.swift
@@ -1,39 +1,0 @@
-// ContentUnavailableView.swift
-//
-// Copyright 2025 FOS Computer Services, LLC
-//
-// Licensed under the Apache License, Version 2.0 (the  License);
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import FOSFoundation
-#if canImport(SwiftUI)
-import SwiftUI
-
-public extension ContentUnavailableView where Label == SwiftUI.Label<Text, Image>, Description == Text?, Actions == EmptyView {
-    /// Displays ``Localizable`` text in the UI
-    ///
-    /// - Parameter localizable: The ``Localizable`` text to be presented
-    /// ```
-    init(
-        _ localizable: some Localizable,
-        systemImage: String,
-        description: Text? = nil,
-        defaultValue: String? = nil
-    ) {
-        self.init(
-            localizable.defaultedLocalizedString(defaultValue: defaultValue),
-            systemImage: systemImage,
-            description: description
-        )
-    }
-}
-#endif
