+++ Sources/FOSMVVM/SwiftUI Support/LocalizableViews.swift
@@ -1,4 +1,4 @@
-// Text.swift
+// LocalizableViews.swift
 //
 // Copyright 2024 FOS Computer Services, LLC
 //
@@ -19,6 +19,12 @@ import FOSFoundation
 import SwiftUI
 
 public extension Text {
+    // Hand-written, not swept: Apple's LocalizedStringKey `Text` init is
+    // `init(_:tableName:bundle:comment:)` — its bundle-lookup parameters have no
+    // meaning for pre-localized strings, so the sweep honestly rejects it
+    // (no-delegate-target). This init is FOSMVVM's own contract, not a mechanical
+    // mirror of an Apple overload.
+
     /// Displays ``Localizable`` text in the UI
     ///
     /// - Parameter localizable: The ``Localizable`` text to be presented
@@ -42,12 +48,82 @@ public extension Text {
         self.init(localizable.defaultedLocalizedString(defaultValue: defaultValue))
     }
 
+    /// Displays optional ``Localizable`` text in the UI
+    ///
+    /// When *localizable* is `nil`, *defaultValue* is rendered instead;
+    /// when that is also `nil`, an empty string is rendered.
+    ///
+    /// ## Example
+    ///
+    /// ```swift
+    /// public struct MyViewModel: ViewModel {
+    ///   public var subtitle: LocalizableString?
+    /// }
+    ///
+    /// struct LandingPageView: ViewModelView {
+    ///   let viewModel: LandingPageViewModel
+    ///
+    ///   var body: some View {
+    ///     Text(viewModel.subtitle, defaultValue: "—")
+    ///   }
+    /// }
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - localizable: The ``Localizable`` text to be presented, if any
+    ///   - defaultValue: Fallback text used when *localizable* is `nil`
+    ///     or its localization did not complete
     init(_ localizable: (some Localizable)?, defaultValue: String? = nil) {
         self.init(localizable?.defaultedLocalizedString(defaultValue: defaultValue) ?? defaultValue ?? "")
     }
 }
 
+public extension LabeledContent where Label == Text, Content == Text {
+    // Hand-written, not swept: Apple's LocalizedStringKey `LabeledContent.init(_:value:)`
+    // has no mechanically-matchable String/Text sibling — the sweep honestly rejects it
+    // (no-delegate-target). This overload is FOSMVVM's own contract, preserved by hand.
+
+    /// Creates a labeled content component with a title generated from a Localizable
+    /// and a string value.
+    ///
+    /// ## Example
+    ///
+    /// ```swift
+    /// LabeledContent(viewModel.balanceLabel, value: account.balance)
+    /// ```
+    ///
+    /// - Parameters:
+    ///    - localizable: A Localizable used as the label.
+    ///    - defaultValue: An optional default string if the localization fails.
+    ///    - value: The string value to display.
+    nonisolated init(
+        _ localizable: some Localizable,
+        defaultValue: String? = nil,
+        value: String
+    ) {
+        self.init {
+            Text(value)
+        } label: {
+            Text(localizable.defaultedLocalizedString(defaultValue: defaultValue))
+        }
+    }
+}
+
 public extension Localizable {
+    /// A view that resolves and displays this ``Localizable`` on the client
+    ///
+    /// Reach for `text` when the value's localization is still pending
+    /// (`localizationStatus == .localizationPending`) — for example, a stubbed
+    /// or locally-constructed value that never round-tripped through the
+    /// server. The view resolves the value on the client via
+    /// ``MVVMEnvironment``'s localization store. For values already localized
+    /// by the server, use `Text(localizable)` instead.
+    ///
+    /// ## Example
+    ///
+    /// ```swift
+    /// viewModel.pageTitle.text
+    /// ```
     var text: some View {
         LocalizableResolverView(localizable: self)
     }
