+++ Sources/FOSMVVM/SwiftUI Support/SweepCoverage.md
@@ -1,0 +1,74 @@
+# Sweep Coverage
+
+GENERATED FILE — DO NOT EDIT.
+Generated by `scripts/localizable-overload-sweep.swift`.
+
+- SDKs: macosx 26.5 | iphoneos 26.5 | appletvos 26.5 | watchos 26.5 | xros 26.5
+- Xcode 26.5 (informational)
+- Regenerate: `swift scripts/localizable-overload-sweep.swift`
+
+## Summary
+
+- Swept candidate rows (pre-union, across all platform graphs): 1021
+- Unique APIs (post-union): 272
+- Generated overloads: 253
+  - policy `direct` (String sibling): 219
+  - policy `text-verbatim` (Text sibling): 34
+- Rejected — `deprecated`: 18
+- Rejected — `no-delegate-target`: 14
+- Rejected — `unrecognized-shape`: 5
+
+## Rejected — deprecated
+
+- `MenuButton.init(_:content:)` — platforms: macosx
+- `NavigationLink.init(_:destination:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `NavigationLink.init(_:destination:isActive:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `NavigationLink.init(_:destination:tag:selection:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `NavigationLink.init(_:isActive:destination:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `NavigationLink.init(_:tag:selection:destination:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `SecureField.init(_:text:onCommit:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:text:onCommit:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:text:onEditingChanged:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:text:onEditingChanged:onCommit:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:value:formatter:onCommit:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:value:formatter:onEditingChanged:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TextField.init(_:value:formatter:onEditingChanged:onCommit:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `View.navigationBarTitle(_:)` — platforms: appletvos, iphoneos, watchos, xros
+- `View.navigationBarTitle(_:displayMode:)` — platforms: iphoneos, xros
+- `View.searchable(text:placement:prompt:suggestions:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `WindowGroup.init(_:content:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `WindowGroup.init(_:id:content:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+
+## Rejected — no-delegate-target
+
+- `AccessibilityCustomContentKey.init(_:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `AccessibilityRotorEntry.init(_:textRange:prepare:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `LabeledContent.init(_:value:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `LocalizedStringKey.init(stringInterpolation:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `Menu.init(_:image:content:primaryAction:)` — platforms: appletvos, iphoneos, macosx, xros
+- `Menu.init(_:systemImage:content:primaryAction:)` — platforms: appletvos, iphoneos, macosx, xros
+- `ModifiedContent.accessibilityCustomContent(_:_:importance:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `ModifiedContent.accessibilityInputLabels(_:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `ModifiedContent.accessibilityInputLabels(_:isEnabled:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `TabContent.accessibilityInputLabels(_:isEnabled:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `Text.init(_:tableName:bundle:comment:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `View.accessibilityCustomContent(_:_:importance:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `View.accessibilityInputLabels(_:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+- `View.accessibilityInputLabels(_:isEnabled:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+
+## Rejected — unrecognized-shape
+
+- `ModifiedContent.accessibilityCustomContent(_:_:importance:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+  - inserted parameter name collides with an existing one — declaration: nonisolated func accessibilityCustomContent(_ labelKey: LocalizedStringKey, _ valueKey: LocalizedStringKey, importance: AXCustomContent.Importance = .default) -> ModifiedContent<Content, Modifier>
+- `View.accessibilityCustomContent(_:_:importance:)` — platforms: appletvos, iphoneos, macosx, watchos, xros
+  - inserted parameter name collides with an existing one — declaration: nonisolated func accessibilityCustomContent(_ labelKey: LocalizedStringKey, _ valueKey: LocalizedStringKey, importance: AXCustomContent.Importance = .default) -> ModifiedContent<Self, AccessibilityAttachmentModifier>
+- `View.badge(_:)` — platforms: iphoneos, macosx, xros
+  - LocalizedStringKey slot '_' is not a plain single-fragment type — declaration: nonisolated func badge(_ key: LocalizedStringKey?) -> some View
+- `WindowGroup.init(_:for:content:defaultValue:)` — platforms: iphoneos, macosx, xros
+  - inserted parameter name collides with an existing one — declaration: nonisolated init<D, C>(_ titleKey: LocalizedStringKey, for type: D.Type = D.self, @ViewBuilder content: @escaping (Binding<D>) -> C, defaultValue: @escaping () -> D) where Content == PresentedWindowContent<D, C>, D : Decodable, D : Encodable, D : Hashable, C : View
+- `WindowGroup.init(_:id:for:content:defaultValue:)` — platforms: iphoneos, macosx, xros
+  - inserted parameter name collides with an existing one — declaration: nonisolated init<D, C>(_ titleKey: LocalizedStringKey, id: String, for type: D.Type = D.self, @ViewBuilder content: @escaping (Binding<D>) -> C, defaultValue: @escaping () -> D) where Content == PresentedWindowContent<D, C>, D : Decodable, D : Encodable, D : Hashable, C : View
+
+## Beta-tier (informational)
+
+None.
