+++ Sources/FOSMVVM/SwiftUI Support/View.swift
@@ -35,27 +35,6 @@ extension View {
 }
 
 public extension View {
-    /// Configures the view's title for purposes of navigation,
-    /// using a localized string.
-    ///
-    /// A view's navigation title is used to visually display
-    /// the current navigation state of an interface.
-    /// On iOS and watchOS, when a view is navigated to inside
-    /// of a navigation view, that view's title is displayed
-    /// in the navigation bar. On iPadOS, the primary destination's
-    /// navigation title is reflected as the window's title in the
-    /// App Switcher. Similarly on macOS, the primary destination's title
-    /// is used as the window title in the titlebar, Windows menu
-    /// and Mission Control.
-    ///
-    /// Refer to the <doc:Configure-Your-Apps-Navigation-Titles> article
-    /// for more information on navigation title modifiers.
-    ///
-    /// - Parameter titleKey: The key to a localized string to display.
-    func navigationTitle(_ localizable: some Localizable) -> some View {
-        navigationTitle(localizable.defaultedLocalizedString())
-    }
-
     /// Indicates when a ``ViewModel`` binding might be out of date
     ///
     /// To invalidate a bound ``ViewModel``, provide a Binding<Bool> that returns
